+++ packages/types/src/api/params.ts
@@ -54,5 +54,12 @@ export enum ApiParams {
 	enabled = "enabled",
 	headHash = "headHash",
 	behavioralData = "behavioralData",
+	decisionMachineScope = "decisionMachineScope",
+	decisionMachineRuntime = "decisionMachineRuntime",
+	decisionMachineSource = "decisionMachineSource",
+	decisionMachineLanguage = "decisionMachineLanguage",
+	decisionMachineName = "decisionMachineName",
+	decisionMachineVersion = "decisionMachineVersion",
+	decisionMachineCaptchaType = "decisionMachineCaptchaType",
 	salt = "salt",
 }